+++ display.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"github.com/davidscholberg/go-durationfmt"
 	"github.com/goodsign/monday"
 	"github.com/julienschmidt/httprouter"
 	"github.com/petrjahoda/database"
@@ -18,17 +17,21 @@ type LcdWorkplaces struct {
 	ProductionPercent string
 	DowntimePercent   string
 	OfflinePercent    string
+	ProductionColor   string
+	DowntimeColor     string
+	PowerOffColor     string
 	Time              string
 	Version           string
 }
 
 type LcdWorkplace struct {
-	StateColor string
-	Name       string
-	User       string
-	Order      string
-	Downtime   string
-	Duration   string
+	StateColor      string
+	StateColorStyle string
+	Name            string
+	User            string
+	Order           string
+	Downtime        string
+	Duration        string
 }
 
 func display1(writer http.ResponseWriter, _ *http.Request, _ httprouter.Params) {
@@ -44,9 +47,13 @@ func display1(writer http.ResponseWriter, _ *http.Request, _ httprouter.Params) 
 		logError("HTML", "Problem opening database: "+err.Error())
 	}
 	db.Order("Name asc").Find(&workplaces)
+	location := downloadActualLocation()
+	loc, _ := time.LoadLocation(location)
 	for _, workplace := range workplaces {
 		stateRecord := database.StateRecord{}
 		db.Where("workplace_id = ?", workplace.ID).Last(&stateRecord)
+		state := database.State{}
+		db.Where("id = ?", stateRecord.StateID).Find(&state)
 		orderRecord := database.OrderRecord{}
 		db.Where("workplace_id = ?", workplace.ID).Where("date_time_end is null").Last(&orderRecord)
 		workplaceHasOpenOrder := orderRecord.ID > 0
@@ -65,25 +72,20 @@ func display1(writer http.ResponseWriter, _ *http.Request, _ httprouter.Params) 
 			userName = user.FirstName + " " + user.SecondName
 		}
 		color := "green"
-		switch stateRecord.StateID {
-		case 1:
-			color = "green"
-		case 2:
-			color = "orange"
-		case 3:
-			color = "red"
-		}
-		duration, err := durationfmt.Format(time.Now().Sub(stateRecord.DateTimeStart), "%dd %hh %mm")
+		colorStyle := "background: " + state.Color
+
+		duration := time.Now().In(loc).Sub(stateRecord.DateTimeStart.In(loc)).Round(1 * time.Second).String()
 		if err != nil {
 			logError(workplace.Name, "Problem parsing datetime: "+err.Error())
 		}
 		logInfo(workplace.Name, duration)
 		logInfo(workplace.Name, "Workplace color: "+color+", order: "+order.Name+", downtime: "+downtime.Name+", user: "+userName)
 		logInfo("HTML", "Adding workplace: "+workplace.Name)
-		lcdWorkplace := LcdWorkplace{Name: workplace.Name, User: userName, StateColor: color, Duration: duration, Downtime: downtime.Name, Order: order.Name}
+		lcdWorkplace := LcdWorkplace{Name: workplace.Name, User: userName, StateColor: color, Duration: duration, Downtime: downtime.Name, Order: order.Name, StateColorStyle: colorStyle}
 		lcdWorkplaces.LcdWorkplaces = append(lcdWorkplaces.LcdWorkplaces, lcdWorkplace)
 	}
-
+	var states []database.State
+	db.Find(&states)
 	db.Find(&workplaces)
 	production := 0
 	downtime := 0
@@ -116,17 +118,25 @@ func display1(writer http.ResponseWriter, _ *http.Request, _ httprouter.Params) 
 	lcdWorkplaces.ProductionPercent = "Production " + strconv.Itoa(productionPercent) + "%"
 	lcdWorkplaces.DowntimePercent = "Downtime " + strconv.Itoa(downtimePercent) + "%"
 	lcdWorkplaces.OfflinePercent = "Poweroff " + strconv.Itoa(offlinePercent) + "%"
+	for _, state := range states {
+		switch state.Name {
+		case "Production":
+			lcdWorkplaces.ProductionColor = state.Color
+		case "Downtime":
+			lcdWorkplaces.DowntimeColor = state.Color
+		case "Poweroff":
+			lcdWorkplaces.PowerOffColor = state.Color
+		}
+	}
 	lcdWorkplaces.Version = "version: " + version
-
 	timezone := readTimeZoneFromDatabase()
 	if timezone == "" {
 		timezone = readTimeZoneFromDatabase()
 	}
-	location, err := time.LoadLocation(timezone)
 	if err != nil {
 		logError("MAIN", "Problem loading location: "+timezone)
 	} else {
-		lcdWorkplaces.Time = monday.Format(time.Now().In(location), "Monday, 2. January 2006, 15:04:05", monday.LocaleEnUS)
+		lcdWorkplaces.Time = monday.Format(time.Now().In(loc), "Monday, 2. January 2006, 15:04:05", monday.LocaleEnUS)
 	}
 	_ = tmpl.Execute(writer, lcdWorkplaces)
 	logInfo("HTML", "Display 1 process ended in "+time.Since(timer).String())