+++ CHANGELOG.md
@@ -18,6 +18,7 @@ adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 
 ### Changed
 - simplified streaming
+- proper color from database
 
 ## [2021.2.1.1] - 2021-04-01